refactor: вынесла вьюхолдер в отдельный файл после ревью
